Add files via upload
This class was enjoyable simply due to the practical application and having the ability to work with the actual tools provided. I do think that through the course there was MUCH filler knowledge that simply went in one ear and out the other but I do believe that I gained a good bit of understanding about databases that I hadn't had previously.